@@ -1,6 +1,6 @@
-module.exports = function (api) {
-  api.cache(true);
+export default function (api) {
+  api.cache(true)
   return {
-    presets: ["babel-preset-expo"],
-  };
-};
+    presets: ['babel-preset-expo'],
+  }
+}
