@@ -1,9 +1,9 @@
-import * as WebBrowser from "expo-web-browser";
-import { StyleSheet, TouchableOpacity } from "react-native";
+import * as WebBrowser from 'expo-web-browser'
+import { StyleSheet, TouchableOpacity } from 'react-native'
 
-import Colors from "../constants/Colors";
-import { MonoText } from "./StyledText";
-import { Text, View } from "./Themed";
+import Colors from '../constants/Colors'
+import { MonoText } from './StyledText'
+import { Text, View } from './Themed'
 
 export default function EditScreenInfo({ path }: { path: string }) {
   return (
@@ -44,18 +44,18 @@ export default function EditScreenInfo({ path }: { path: string }) {
         </TouchableOpacity>
       </View>
     </View>
-  );
+  )
 }
 
 function handleHelpPress() {
   WebBrowser.openBrowserAsync(
-    "https://docs.expo.io/get-started/create-a-new-app/#opening-the-app-on-your-phonetablet"
-  );
+    'https://docs.expo.io/get-started/create-a-new-app/#opening-the-app-on-your-phonetablet',
+  )
 }
 
 const styles = StyleSheet.create({
   getStartedContainer: {
-    alignItems: "center",
+    alignItems: 'center',
     marginHorizontal: 50,
   },
   homeScreenFilename: {
@@ -68,17 +68,17 @@ const styles = StyleSheet.create({
   getStartedText: {
     fontSize: 17,
     lineHeight: 24,
-    textAlign: "center",
+    textAlign: 'center',
   },
   helpContainer: {
     marginTop: 15,
     marginHorizontal: 20,
-    alignItems: "center",
+    alignItems: 'center',
   },
   helpLink: {
     paddingVertical: 15,
   },
   helpLinkText: {
-    textAlign: "center",
+    textAlign: 'center',
   },
-});
+})
