@@ -1,7 +1,5 @@
-import { Text, TextProps } from "./Themed";
+import { Text, TextProps } from './Themed'
 
 export function MonoText(props: TextProps) {
-  return (
-    <Text {...props} style={[props.style, { fontFamily: "space-mono" }]} />
-  );
+  return <Text {...props} style={[props.style, { fontFamily: 'space-mono' }]} />
 }
