@@ -1,10 +1,10 @@
-import * as React from "react";
-import renderer from "react-test-renderer";
+import * as React from 'react'
+import renderer from 'react-test-renderer'
 
-import { MonoText } from "../StyledText";
+import { MonoText } from '../StyledText'
 
-it(`renders correctly`, () => {
-  const tree = renderer.create(<MonoText>Snapshot test!</MonoText>).toJSON();
+it('renders correctly', () => {
+  const tree = renderer.create(<MonoText>Snapshot test!</MonoText>).toJSON()
 
-  expect(tree).toMatchSnapshot();
-});
+  expect(tree).toMatchSnapshot()
+})
