@@ -1,19 +1,19 @@
-const tintColorLight = "#2f95dc";
-const tintColorDark = "#fff";
+const tintColorLight = '#2f95dc'
+const tintColorDark = '#fff'
 
 export default {
   light: {
-    text: "#000",
-    background: "#fff",
+    text: '#000',
+    background: '#fff',
     tint: tintColorLight,
-    tabIconDefault: "#ccc",
+    tabIconDefault: '#ccc',
     tabIconSelected: tintColorLight,
   },
   dark: {
-    text: "#fff",
-    background: "#000",
+    text: '#fff',
+    background: '#000',
     tint: tintColorDark,
-    tabIconDefault: "#ccc",
+    tabIconDefault: '#ccc',
     tabIconSelected: tintColorDark,
   },
-};
+}
