@@ -1,7 +1,7 @@
-import { Dimensions } from "react-native";
+import { Dimensions } from 'react-native'
 
-const width = Dimensions.get("window").width;
-const height = Dimensions.get("window").height;
+const width = Dimensions.get('window').width
+const height = Dimensions.get('window').height
 
 export default {
   window: {
@@ -9,4 +9,4 @@ export default {
     height,
   },
   isSmallDevice: width < 375,
-};
+}
