@@ -1,33 +1,33 @@
-import { FontAwesome } from "@expo/vector-icons";
-import * as Font from "expo-font";
-import * as SplashScreen from "expo-splash-screen";
-import { useEffect, useState } from "react";
+import { FontAwesome } from '@expo/vector-icons'
+import * as Font from 'expo-font'
+import * as SplashScreen from 'expo-splash-screen'
+import { useEffect, useState } from 'react'
 
 export default function useCachedResources() {
-  const [isLoadingComplete, setLoadingComplete] = useState(false);
+  const [isLoadingComplete, setLoadingComplete] = useState(false)
 
   // Load any resources or data that we need prior to rendering the app
   useEffect(() => {
     async function loadResourcesAndDataAsync() {
       try {
-        SplashScreen.preventAutoHideAsync();
+        SplashScreen.preventAutoHideAsync()
 
         // Load fonts
         await Font.loadAsync({
           ...FontAwesome.font,
-          "space-mono": require("../assets/fonts/SpaceMono-Regular.ttf"),
-        });
+          'space-mono': require('../assets/fonts/SpaceMono-Regular.ttf'),
+        })
       } catch (e) {
         // We might want to provide this error information to an error reporting service
-        console.warn(e);
+        console.warn(e)
       } finally {
-        setLoadingComplete(true);
-        SplashScreen.hideAsync();
+        setLoadingComplete(true)
+        SplashScreen.hideAsync()
       }
     }
 
-    loadResourcesAndDataAsync();
-  }, []);
+    loadResourcesAndDataAsync()
+  }, [])
 
-  return isLoadingComplete;
+  return isLoadingComplete
 }
